size_t changed for imgno
git-svn-id: file:///home/roberto/Git/svn@7128 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 2cfbba65eb69c3f41c3e116fd1dffc3adf0783f4

+++ trunk/xmipp/libraries/reconstruction/angular_projection_matching.cpp
@@ -917,7 +917,7 @@ void ProgAngularProjectionMatching::processSomeImages(int * my_images, double * 
 
 }
 
-void ProgAngularProjectionMatching::getCurrentImage(int imgno, Image<double> &img)
+void ProgAngularProjectionMatching::getCurrentImage(size_t imgno, Image<double> &img)
 {
 
     FileName fn_img;
